extracting dialog component and moving shared interfaces to data file
