+++ jupyterlab_gcedetails/src/data.ts
@@ -1,0 +1,143 @@
+/**
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as csstips from 'csstips';
+import { stylesheet } from 'typestyle';
+
+interface MachineType {
+    name: string;
+    description: string;
+}
+  
+interface Instance {
+  attributes: {
+    framework: string;
+    title: string;
+    version: string;
+  };
+  cpuPlatform: string;
+  id: number;
+	image: string;
+	machineType: MachineType;
+  name: string;
+  zone: string;
+}
+  
+interface Project {
+  numericProjectId: number;
+  projectId: string;
+}
+  
+interface Utilization {
+  cpu: number;
+  memory: number;
+}
+  
+interface Gpu {
+  name: string;
+  driver_version: string;
+  cuda_version: string;
+  gpu: number;
+  memory: number;
+  temperature: number;
+}
+  
+export interface Details {
+  instance: Instance;
+  project: Project;
+  utilization: Utilization;
+  gpu: Gpu;
+}
+  
+interface AttributeMapper {
+  label: string;
+  mapper: (details: Details) => string;
+}
+
+// Displayable virtual machine detail attributes
+export const MAPPED_ATTRIBUTES: AttributeMapper[] = [
+  { label: 'VM Name', mapper: (details: Details) => details.instance.name },
+  { label: 'Project', mapper: (details: Details) => details.project.projectId },
+  {
+    label: 'Framework',
+    mapper: (details: Details) => details.instance.attributes.framework,
+  },
+  {
+    label: 'Machine Type',
+    mapper: (details: Details) =>
+      `${details.instance.machineType.description} (${details.instance.machineType.name})`,
+  },
+];
+
+export const REFRESHABLE_MAPPED_ATTRIBUTES = [
+  {
+    label: 'CPU Utilization',
+    mapper: (details: Details) => `CPU: ${details.utilization.cpu.toFixed(1)}%`,
+  },
+  {
+    label: 'Memory Utilization',
+    mapper: (details: Details) =>
+      `Memory: ${details.utilization.memory.toFixed(1)}%`,
+  },
+  {
+    label: 'GPU Utilization',
+    mapper: (details: Details) => {
+      if (!details.gpu.name) {
+        return 'No GPUs';
+      }
+      return `GPU: ${details.gpu.name} - ${details.gpu.gpu.toFixed(1)}%`;
+    },
+  },
+];
+
+MAPPED_ATTRIBUTES.push(...REFRESHABLE_MAPPED_ATTRIBUTES);
+
+/* Class names applied to the component. Exported for test selectors. */
+export const STYLES = stylesheet({
+  container: {
+    color: 'var(--jp-ui-font-color1)',
+    cursor: 'pointer',
+    fontFamily: 'var(--jp-ui-font-family)',
+    fontSize: 'var(--jp-ui-font-size1, 13px)',
+    lineHeight: '24px',
+    alignItems: 'center',
+    ...csstips.horizontal,
+  },
+  attribute: {
+    marginRight: '4px',
+  },
+  dt: {
+    display: 'table-cell',
+    fontWeight: 'bold',
+    lineHeight: '20px',
+    padding: '2px',
+    verticalAlign: 'top',
+  },
+  dd: {
+    padding: '2px 2px 2px 24px',
+    verticalAlign: 'top',
+  },
+  icon: {
+    display: 'inline-block',
+    height: '18px',
+    marginRight: '4px',
+    width: '18px',
+  },
+  listRow: {
+    display: 'table-row',
+    boxShadow: 'inset 0 -1px 0 0 var(--jp-border-color0)',
+  },
+});
